+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add billing title field to the checkout page / my account > my address page
 * Sync Mailchimp user details with merge tags: PRENOM, NOM, CIV
 * Adapt to the processed option in vouchers
+* Compatibility with WooCommerce 3.3.0
 
 ### 1.0.10: January 10th, 2018
 * Sort order actions: view in first position